+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF2RJ6EATVXSZC7TGS74M9J2"
+        "01KF392YKGA6KMWGJS7ZGN6GZQ"
     )
     r"""The unique identifier of the data source"""