+++ packages/orq-rc/RELEASES.md
@@ -4729,3 +4729,13 @@ Based on:
 - [python v4.2.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.35 - packages/orq-rc
+
+## 2026-01-16 11:31:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.686.0 (2.796.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.36] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.36] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.36 - packages/orq-rc